third_party/blink/web_tests/virtual/web-components-v0-disabled/shadow-dom/range-surround-contents-expected.txt

CONSOLE ERROR: line 29: Uncaught TypeError: parent.createShadowRoot is not a function
This is a testharness.js-based test.
FAIL range-surround-contents Uncaught TypeError: parent.createShadowRoot is not a function
Harness: the test ran to completion.
